feat: Added Unity Networking
This commit adds client RPC networking to the Unity project, allowing clients to communicate with the server/host and trigger remote procedures.

Changes Made:
- Implemented the Unity Networking package
- Implemented the Client Network Transform on the player object
- Added a new NetworkManager object to the scene and configured it to use the client-side networking script
- Modified the relevant game objects and scripts to use the new client RPC functions instead of the previous method for handling networking
- Added Helper UI for testing with the networking
- Added the Odin Validator Asset

This commit improves the overall functionality of the Unity project by adding multiplayer support, although it should be noted that the current implementation trusts the clients and may be unsafe. Consider revising this implementation at some point to ensure greater security.