feat(Player): Added Health and Fixed Data Sync
This commit adds a health bar that follows the player, and can be seen by anyone. The health bar tracks the health of the player, and is updated when they take damage. There was also an issue with data not syncing across the sever. Specifically, it did not sync the ID of the player, and the Damage Multiplier from the player.
Also, detecting when a bullet hits a player is done by  using OnTriggerEnter, but it verified by making sure that the ID of the original object is not equal to the object it collided with.