feat(Scenes): Added Lobby
This commit adds an authentications scene, and a lobby scene. The Authentication scene will log a player in anonymously, and then send them to the lobby scene. This allows for players to get playing quickly.
The lobby scene lets you browse lobbies, or create one if you would like. If you do either, it will put you in the play scene that that lobby is connected to.