feat(Player): Add Gamemode Infrastructure
This commit adds the necessary infrastructure to be able to implement gamemodes in the future. This is done by creating a single class for the player to control gamemode specific logic. This commit also makes some changes to the Constants class for the lobby to make it easier to work with. Specifically, the GameTypes field is now an enum, rather than an array. There is also a field to convert the enum into an array when necessary.

Changes Made:
- Added Player Gamemodes class
- Changed Constants class
- Made it so that the player only respawns in certain gamemodes