feat(Player): Add Bots
This commit adds Ai powered bots that can follow the player and pathfind around objects. This is very important to the game, because there may not always be players online to play against. The bots are implemented using NavAgents, and baking a mesh with all the obstacles.
This commit also adds some scripts pertaining to a leaderboard, however, it did not work, and therefor is not important. I may make use of it later however.

Changes Made:
- Added Bots