feat(Bots): Add Bot Shooting
This commit gives gits the ability to shoot, along with fixing a bunch of other issues. To do this, a BotShooting Script was added. This script creates a raycast directly Infront of it, and if it sees a player, it will shoot after a bit of time. The bot's navigation is also not controlled by distance anymore. Instead it is controlled by line of sight. If the bot can see a player, it stops. If it cannot, then it will continue.
Some issues were also fixed in this commit. There were a bunch of issues due to the use of transform.root, which did not work with bots because they are nested. Another issue that was happening was that every hit would be registered twice on the player. This was because each bullet was spawned locally and on the network, which was nice because it was responsive, but it also created two bullets. I thought this would not be an issue because the bullets were local, but because the player's health was not local, it did not work. To fix this, I just made the bullets multiplayer, and spawned over the network.

Changes Made:
- Bots can now shoot
- Fixed Bugs