feat(): Add Toon Shading
This commit adds toon shading to the project. This gives the game a more stylized look, and overall, adds to its uniqueness. This commit also fixes a few bugs with the bots, namely it fixes a glitch where bots would not die, and it fixes the player not receiving points for damaging or killing enemies.

Changes Made:
- Added Toon Shading
- Fixed Bugs with Bots